# Reduce ingester memory by using TSDB's index