**What this PR does**:

`RefCache` is doing almost exactly the same work as TSDB, with a sharded map of series by labels.
prometheus/prometheus#8600 and prometheus/prometheus#8641 add a method to `TSDB.Appender` which allows us to save building a parallel cache, reducing ingester heap by about 20%.

We depend on values from `GetRef()` being valid to retain elsewhere.

Also added a benchmark, broadly copied from `BenchmarkIngesterPush()`. 

```
name              old time/op    new time/op    delta
IngesterV2Push-4     448ms ± 3%     388ms ± 1%  -13.42%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op   delta
IngesterV2Push-4    29.1MB ± 1%    27.8MB ± 0%   -4.57%  (p=0.016 n=5+4)

name              old allocs/op  new allocs/op  delta
IngesterV2Push-4     64.1k ± 0%     53.5k ± 0%  -16.52%  (p=0.008 n=5+5)
```

Tests show the memory reduction is around 20% when TSDB's index has been created from the WAL, in which case `RefCache` has a copy of all the labels for series that have received more data after restart.
Not much CPU reduction overall, since series are typically looked up many more times than they are added.

**Checklist**
- [x] Tests updated
- NA Documentation added
- [x] `CHANGELOG.md` updated