@pstibrany 2021-03-15
This looks like nice improvement, but I think we should get the change into Prometheus first, to avoid using forked version.