@bboreham 2021-03-15
Yeah I thought I would get some traction here before opening the Prometheus PR. 
E.g. perhaps I missed some reason for having RefCache.