@bboreham 2021-03-20
Updated after Prometheus PR merged.