@bboreham 2021-03-25
I'm running this branch in production now; Go heap is 20% smaller as predicted.